+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * made `readenv.redenv.load` "private" (`readenv.readenv._load`)
 * rename `readenv.readenv.loads` to `readenv.readenv.load`
 * export `readenv.readenv.load` as `readenv.load`
+* added `readenv.env.get`, `readenv.env.set`, and `readenv.env.setdefault`,
+  and exported to `readenv`
 
 ## 0.1.1